@@ -1,23 +1,118 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
+import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { MyComponentComponent } from './invoice/my-component/my-component.component';
-import { TestCompComponent } from './testComp/testComp.component';
-import { CompanyComponent } from './company/company.component';
-import {HttpClientModule} from '@angular/common/http';
-
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { CompanyListComponent } from './company-list/company-list.component';
+import { HttpClientModule } from '@angular/common/http';
+import {DragDropModule} from '@angular/cdk/drag-drop';
+import {ScrollingModule} from '@angular/cdk/scrolling';
+import {CdkTableModule} from '@angular/cdk/table';
+import {CdkTreeModule} from '@angular/cdk/tree';
+import {
+   MatAutocompleteModule,
+   MatBadgeModule,
+   MatBottomSheetModule,
+   MatButtonModule,
+   MatButtonToggleModule,
+   MatCardModule,
+   MatCheckboxModule,
+   MatChipsModule,
+   MatDatepickerModule,
+   MatDialogModule,
+   MatDividerModule,
+   MatExpansionModule,
+   MatGridListModule,
+   MatIconModule,
+   MatInputModule,
+   MatListModule,
+   MatMenuModule,
+   MatNativeDateModule,
+   MatPaginatorModule,
+   MatProgressBarModule,
+   MatProgressSpinnerModule,
+   MatRadioModule,
+   MatRippleModule,
+   MatSelectModule,
+   MatSidenavModule,
+   MatSliderModule,
+   MatSlideToggleModule,
+   MatSnackBarModule,
+   MatSortModule,
+   MatStepperModule,
+   MatTableModule,
+   MatTabsModule,
+   MatToolbarModule,
+   MatTooltipModule,
+   MatTreeModule,
+ } from '@angular/material';
+import { CompanyList2Component } from './company-list2/company-list2.component';
+import { AitCloudNavComponent } from './ait-cloud-nav/ait-cloud-nav.component';
+import { LayoutModule } from '@angular/cdk/layout';
 
 @NgModule({
    declarations: [
       AppComponent,
-      MyComponentComponent,
-      TestCompComponent,
-      CompanyComponent,
+      CompanyListComponent,
+      CompanyList2Component,
+      AitCloudNavComponent
    ],
    imports: [
       BrowserModule,
-      HttpClientModule
+      AppRoutingModule,
+      BrowserAnimationsModule,
+      HttpClientModule,
+      MatTableModule,
+      MatPaginatorModule,
+      MatSortModule,
+      LayoutModule,
+      MatToolbarModule,
+      MatButtonModule,
+      MatSidenavModule,
+      MatIconModule,
+      MatListModule
    ],
+   exports: [
+      CdkTableModule,
+      CdkTreeModule,
+      DragDropModule,
+      MatAutocompleteModule,
+      MatBadgeModule,
+      MatBottomSheetModule,
+      MatButtonModule,
+      MatButtonToggleModule,
+      MatCardModule,
+      MatCheckboxModule,
+      MatChipsModule,
+      MatStepperModule,
+      MatDatepickerModule,
+      MatDialogModule,
+      MatDividerModule,
+      MatExpansionModule,
+      MatGridListModule,
+      MatIconModule,
+      MatInputModule,
+      MatListModule,
+      MatMenuModule,
+      MatNativeDateModule,
+      MatPaginatorModule,
+      MatProgressBarModule,
+      MatProgressSpinnerModule,
+      MatRadioModule,
+      MatRippleModule,
+      MatSelectModule,
+      MatSidenavModule,
+      MatSliderModule,
+      MatSlideToggleModule,
+      MatSnackBarModule,
+      MatSortModule,
+      MatTableModule,
+      MatTabsModule,
+      MatToolbarModule,
+      MatTooltipModule,
+      MatTreeModule,
+      ScrollingModule,
+    ],
    providers: [],
    bootstrap: [
       AppComponent
